requirements.txt
# HYDRA - AI-Native Crypto Perpetual Futures Trading System
# Core Dependencies

# Async & Networking
aiohttp>=3.9.0
websockets>=12.0
httpx>=0.25.0

# Data Processing
numpy>=1.26.0
pandas>=2.1.0
polars>=0.19.0
scipy>=1.11.0

# Machine Learning & Deep Learning
torch>=2.1.0
transformers>=4.35.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.2.0

# Statistical Modeling
statsmodels>=0.14.0
arch>=6.2.0

# LLM Integration
openai>=1.0.0
anthropic>=0.18.0
groq>=0.4.0
tiktoken>=0.5.0

# Technical Analysis
ta>=0.11.0
pandas-ta>=0.3.14b

# Exchange APIs
ccxt>=4.1.0
python-binance>=1.0.19

# On-Chain Data
web3>=6.11.0

# Database & Caching
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0

# Configuration & CLI
pydantic>=2.5.0
pydantic-settings>=2.1.0
typer>=0.9.0
rich>=13.7.0

# Monitoring & Logging
loguru>=0.7.0
prometheus-client>=0.19.0

# Visualization (for analysis)
plotly>=5.18.0
matplotlib>=3.8.0

# Dashboard
streamlit>=1.29.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
orjson>=3.9.0
tenacity>=8.2.0
pytz>=2023.3
